+++ contracts/mocks/AuxHelperTestHarness.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.21;
+pragma solidity 0.8.30;
 
 import '../utility/AuxHelper16.sol';
 import '../utility/AuxHelper32.sol';