+++ contracts/token/opensea/seadrop/ERC721SeaDrop.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.21;
+pragma solidity 0.8.30;
 
 import {ERC721ContractMetadata, ISeaDropTokenContractMetadata} from './ERC721ContractMetadata.sol';