+++ contracts/token/opensea/seadrop/lib/ERC721AConduitPreapproved.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.21;
+pragma solidity 0.8.30;
 
 import { ERC721A } from "erc721a/contracts/ERC721A.sol";