+++ contracts/token/opensea/seadrop/lib/SeaDropStructs.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.21;
+pragma solidity 0.8.30;
 
 /**
  * @notice A struct defining public drop data.